refactor(core): fix packages
