+++ example/compose/build.gradle.kts
@@ -50,7 +50,7 @@ android {
 
 dependencies {
 
-    implementation("com.neoutils.highlight:highlight-compose:2.1.0-snapshot3")
+    implementation("com.neoutils.highlight:highlight-compose:2.1.0-snapshot4")
 
     implementation(libs.androidx.core.ktx)
