+++ example/view-based/src/main/java/com/neoutils/highlight/example/view/example/BackgroundSchemeExample.kt
@@ -6,7 +6,7 @@ import androidx.appcompat.widget.AppCompatTextView
 import com.neoutils.highlight.core.Highlight
 import com.neoutils.highlight.core.extension.backgroundColor
 import com.neoutils.highlight.core.extension.textColor
-import com.neoutils.highlight.core.utils.UiColor
+import com.neoutils.highlight.core.util.UiColor
 import com.neoutils.highlight.view.extension.toSpannedString
 
 class BackgroundSchemeExample(
