+++ example/view-based/src/main/java/com/neoutils/highlight/example/view/example/JavaCaseExample.java
@@ -7,8 +7,8 @@ import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 import androidx.appcompat.widget.AppCompatTextView;
 import com.neoutils.highlight.core.Highlight;
-import com.neoutils.highlight.core.utils.Match;
-import com.neoutils.highlight.core.utils.UiColor;
+import com.neoutils.highlight.core.util.Match;
+import com.neoutils.highlight.core.util.UiColor;
 import com.neoutils.highlight.view.extension.HighlightKt;
 import com.neoutils.highlight.core.scheme.TextColorScheme;
 import kotlin.text.Regex;
