+++ highlight/compose/src/commonMain/kotlin/com/neoutils/highlight/compose/scheme/SpanStyleScheme.kt
@@ -2,7 +2,7 @@ package com.neoutils.highlight.compose.scheme
 
 import androidx.compose.ui.text.SpanStyle
 import com.neoutils.highlight.core.Scheme
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.util.Match
 
 /**
  * Only supported in compose.
