+++ highlight/compose/src/commonMain/kotlin/com/neoutils/highlight/compose/scope/SpanStyleScope.kt
@@ -3,7 +3,7 @@ package com.neoutils.highlight.compose.scope
 import androidx.compose.ui.text.SpanStyle
 import com.neoutils.highlight.compose.scheme.SpanStyleScheme
 import com.neoutils.highlight.core.scope.SchemeScope
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.util.Match
 
 class SpanStyleScope internal constructor() :
     SchemeScope<SpanStyle, SpanStyleScheme>() {
