--- highlight/core/src/commonMain/kotlin/com/neoutils/highlight/core/BackgroundColorScheme.kt
@@ -1,9 +1,0 @@
-package com.neoutils.highlight.core
-
-import com.neoutils.highlight.core.utils.Match
-import com.neoutils.highlight.core.utils.UiColor
-
-data class BackgroundColorScheme(
-    override val regex: Regex,
-    override val match: Match<UiColor>,
-) : Scheme<UiColor>
