+++ highlight/core/src/commonMain/kotlin/com/neoutils/highlight/core/scope/SchemeScope.kt
@@ -2,7 +2,7 @@ package com.neoutils.highlight.core.scope
 
 import com.neoutils.highlight.core.Scheme
 import com.neoutils.highlight.core.Scope
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.util.Match
 import org.intellij.lang.annotations.Language
 
 abstract class SchemeScope<T : Any, S : Scheme<*>> : Scope<S>() {
