+++ highlight/core/src/commonMain/kotlin/com/neoutils/highlight/core/scope/TextColorScope.kt
@@ -1,8 +1,8 @@
 package com.neoutils.highlight.core.scope
 
-import com.neoutils.highlight.core.TextColorScheme
-import com.neoutils.highlight.core.utils.Match
-import com.neoutils.highlight.core.utils.UiColor
+import com.neoutils.highlight.core.scheme.TextColorScheme
+import com.neoutils.highlight.core.util.Match
+import com.neoutils.highlight.core.util.UiColor
 
 class TextColorScope internal constructor() :
     SchemeScope<UiColor, TextColorScheme>() {
