+++ highlight/view-based/src/androidMain/kotlin/com/neoutils/highlight/view/extension/Highlight.kt
@@ -7,8 +7,8 @@ import android.text.style.ForegroundColorSpan
 import android.text.style.StyleSpan
 import com.neoutils.highlight.core.Highlight
 import com.neoutils.highlight.core.Scheme
-import com.neoutils.highlight.core.BackgroundColorScheme
-import com.neoutils.highlight.core.TextColorScheme
+import com.neoutils.highlight.core.scheme.BackgroundColorScheme
+import com.neoutils.highlight.core.scheme.TextColorScheme
 import com.neoutils.highlight.view.scheme.SpanScheme
 import com.neoutils.highlight.view.scheme.TextFontScheme
 import com.neoutils.highlight.view.scheme.TextStyleScheme
