+++ highlight/view-based/src/androidMain/kotlin/com/neoutils/highlight/view/extension/UiColor.kt
@@ -1,7 +1,7 @@
 package com.neoutils.highlight.view.extension
 
 import android.graphics.Color
-import com.neoutils.highlight.core.utils.UiColor
+import com.neoutils.highlight.core.util.UiColor
 
 fun UiColor.toIntColor() = when (this) {
     is UiColor.Hex -> Color.parseColor(hex)
