+++ highlight/view-based/src/androidMain/kotlin/com/neoutils/highlight/view/scheme/TextStyleScheme.kt
@@ -1,7 +1,7 @@
 package com.neoutils.highlight.view.scheme
 
 import com.neoutils.highlight.core.Scheme
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.util.Match
 import com.neoutils.highlight.view.util.UiStyle
 
 data class TextStyleScheme(
