+++ highlight/view-based/src/androidMain/kotlin/com/neoutils/highlight/view/scope/SpanScope.kt
@@ -2,7 +2,7 @@ package com.neoutils.highlight.view.scope
 
 import android.text.ParcelableSpan
 import com.neoutils.highlight.core.scope.SchemeScope
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.util.Match
 import com.neoutils.highlight.view.scheme.SpanScheme
 
 class SpanScope internal constructor() :
