+++ highlight/view-based/src/androidMain/kotlin/com/neoutils/highlight/view/scope/TextStyleScope.kt
@@ -1,7 +1,7 @@
 package com.neoutils.highlight.view.scope
 
 import com.neoutils.highlight.core.scope.SchemeScope
-import com.neoutils.highlight.core.utils.Match
+import com.neoutils.highlight.core.util.Match
 import com.neoutils.highlight.view.scheme.TextStyleScheme
 import com.neoutils.highlight.view.util.UiStyle
